+++ remotecommanddispatcher/src/test/java/com/tealium/remotecommanddispatcher/RemoteCommandRequestTest.kt
@@ -1,7 +1,7 @@
 package com.tealium.remotecommanddispatcher
 
+import com.tealium.internal.tagbridge.RemoteCommandRequest
 import com.tealium.remotecommanddispatcher.remotecommands.JsonRemoteCommand
-import com.tealium.remotecommanddispatcher.remotecommands.RemoteCommand
 import junit.framework.Assert
 import org.json.JSONException
 import org.json.JSONObject
@@ -17,7 +17,7 @@ class RemoteCommandRequestTest {
     @Test
     fun webViewRemoteCommandInvalidCorrupted() {
         try {
-            val request = RemoteCommandRequest.tagManagementRequest("tealium://command?request=%7B%22foo%22%3A%22%7D")
+            val request = RemoteCommandRequest("tealium://command?request=%7B%22foo%22%3A%22%7D")
             Assert.fail()
             Assert.assertNull(request)
         } catch (ex: JSONException) {
@@ -27,7 +27,7 @@ class RemoteCommandRequestTest {
     @Test
     fun webViewRemoteCommandInvalidSymbol() {
         try {
-            val request = RemoteCommandRequest.tagManagementRequest("tealium://command?request=%7B%22foo%22%3A%22bar%22%!!")
+            val request = RemoteCommandRequest("tealium://command?request=%7B%22foo%22%3A%22bar%22%!!")
             Assert.fail()
             Assert.assertNull(request)
         } catch (ex: IllegalArgumentException) {
@@ -36,54 +36,57 @@ class RemoteCommandRequestTest {
 
     @Test
     fun webViewRemoteCommandInvalidCommandName() {
-        val request = RemoteCommandRequest.tagManagementRequest("tealium://commandpayload=%7B%7D")
-
-        Assert.assertNull(request)
+        try {
+            val request = RemoteCommandRequest("tealium://commandpayload=%7B%7D")
+            Assert.fail()
+            Assert.assertNull(request)
+        } catch (ex: IllegalArgumentException) {
+        }
     }
 
     @Test
     fun webViewRemoteCommandValidRequest() {
-        val request = RemoteCommandRequest.tagManagementRequest("tealium://command?request=%7B%22foo%22%3A%22bar%22%7D")
+        val request = RemoteCommandRequest("tealium://command?request=%7B%22foo%22%3A%22bar%22%7D")
 
-        Assert.assertEquals("command", request?.commandId)
-        Assert.assertEquals(0, request?.response?.requestPayload?.length())
+        Assert.assertEquals("command", request.commandId)
+        Assert.assertEquals(0, request.response?.requestPayload?.length())
     }
 
     @Test
     fun webViewRemoteCommandValidRequestWithPayload() {
-        val request = RemoteCommandRequest.tagManagementRequest("tealium://task-populated_arg?request=%7B%22payload%22%3A%7B%22foo%22%3A%22bar%22%7D%7D")
+        val request = RemoteCommandRequest("tealium://task-populated_arg?request=%7B%22payload%22%3A%7B%22foo%22%3A%22bar%22%7D%7D")
 
-        Assert.assertEquals("task-populated_arg", request?.commandId)
-        Assert.assertEquals(1, request?.response?.requestPayload?.length())
-        Assert.assertNull(request?.response?.responseId)
+        Assert.assertEquals("task-populated_arg", request.commandId)
+        Assert.assertEquals(1, request.response?.requestPayload?.length())
+        Assert.assertNull(request.response?.id)
     }
 
     @Test
     fun webViewRemoteCommandValidRequestWithId() {
-        val request = RemoteCommandRequest.tagManagementRequest("tealium://command?request=%7B%22config%22%3A%7B%22response_id%22%3A%7B%7D%7D%7D")
+        val request = RemoteCommandRequest("tealium://command?request=%7B%22config%22%3A%7B%22response_id%22%3A%7B%7D%7D%7D")
 
-        Assert.assertEquals("command", request?.commandId)
-        Assert.assertEquals(0, request?.response?.requestPayload?.length())
-        Assert.assertNotNull(request?.response?.responseId)
+        Assert.assertEquals("command", request.commandId)
+        Assert.assertEquals(0, request.response?.requestPayload?.length())
+        Assert.assertNotNull(request.response?.id)
     }
 
 
     @Test
     fun webViewRemoteCommandValidRequestWithResponseJsonArrayIdAsString() {
-        val request = RemoteCommandRequest.tagManagementRequest("tealium://command?request=%7B%22config%22%3A%7B%22response_id%22%3A%5B%5D%7D%7D")
+        val request = RemoteCommandRequest("tealium://command?request=%7B%22config%22%3A%7B%22response_id%22%3A%5B%5D%7D%7D")
 
-        Assert.assertEquals("command", request?.commandId)
-        Assert.assertEquals(0, request?.response?.requestPayload?.length())
-        Assert.assertNotNull(request?.response?.responseId)
+        Assert.assertEquals("command", request.commandId)
+        Assert.assertEquals(0, request.response?.requestPayload?.length())
+        Assert.assertNotNull(request.response?.id)
     }
 
     @Test
     fun webViewRemoteCommandValidRequestWithResponseNumberIdAsString() {
-        val request = RemoteCommandRequest.tagManagementRequest("tealium://command?request=%7B%22config%22%3A%7B%22response_id%22%3A1234%7D%7D")
+        val request = RemoteCommandRequest("tealium://command?request=%7B%22config%22%3A%7B%22response_id%22%3A1234%7D%7D")
 
-        Assert.assertEquals("command", request?.commandId)
-        Assert.assertEquals(0, request?.response?.requestPayload?.length())
-        Assert.assertNotNull(request?.response?.responseId)
+        Assert.assertEquals("command", request.commandId)
+        Assert.assertEquals(0, request.response?.requestPayload?.length())
+        Assert.assertNotNull(request.response?.id)
     }
 
     @Test
@@ -92,7 +95,7 @@ class RemoteCommandRequestTest {
             override fun onInvoke(response: Response) { // do nothing
             }
         }
-        val request = RemoteCommandRequest.jsonRequest(command, JSONObject())
+        val request = RemoteCommandRequest(command.id, JSONObject())
 
         Assert.assertEquals("jsonTest".toLowerCase(Locale.ROOT), request.commandId)
         Assert.assertNotNull(request.payload)
@@ -105,7 +108,7 @@ class RemoteCommandRequestTest {
             override fun onInvoke(response: Response) { // do nothing
             }
         }
-        val request = RemoteCommandRequest.jsonRequest(command, JSONObject())
+        val request = RemoteCommandRequest(command.id, JSONObject())
 
         Assert.assertEquals("testCommand".toLowerCase(Locale.ROOT), request.commandId)
         Assert.assertNotNull(request.payload)
@@ -118,7 +121,7 @@ class RemoteCommandRequestTest {
             override fun onInvoke(response: Response) { // do nothing
             }
         }
-        val request = RemoteCommandRequest.jsonRequest(command, JSONObject())
+        val request = RemoteCommandRequest(command.id, JSONObject())
 
         Assert.assertEquals("testCommand".toLowerCase(Locale.ROOT), request.commandId)
         Assert.assertNotNull(request.payload)