+++ tagmanagementdispatcher/src/test/java/com/tealium/tagmanagementdispatcher/TagManagementDispatcherTest.kt
@@ -148,8 +148,6 @@ class TagManagementDispatcherTest {
         val dispatch = TealiumEvent("test", mapOf("key" to "value"))
         tagManagementDispatcher.onDispatchReady(dispatch)
 
-        verify {
-            mockWebViewLoader.initializeWebView()
-        }
+        coVerify { mockWebViewLoader.initializeWebView() }
     }
 }