+++ tealiumlibrary/src/main/java/com/tealium/core/messaging/DispatchRouter.kt
@@ -3,6 +3,7 @@ package com.tealium.core.messaging
 import com.tealium.core.Collector
 import com.tealium.core.settings.LibrarySettingsManager
 import com.tealium.core.Logger
+import com.tealium.core.Transformer
 import com.tealium.core.consent.ConsentManagementPolicy
 import com.tealium.core.consent.ConsentManager
 import com.tealium.core.consent.ConsentStatus
@@ -18,6 +19,7 @@ import kotlinx.coroutines.launch
 
 internal class DispatchRouter(coroutineDispatcher: CoroutineDispatcher,
                               private val collectors: Set<Collector>,
+                              private val transformers: Set<Transformer>,
                               private val validators: Set<DispatchValidator>,
                               private val dispatchStore: DispatchStorage,
                               private val librarySettingsManager: LibrarySettingsManager,
@@ -43,6 +45,7 @@ internal class DispatchRouter(coroutineDispatcher: CoroutineDispatcher,
         scope.launch(Logger.exceptionHandler) {
             // Collection
             dispatch.addAll(collect())
+            transform(dispatch)
 
             // Validation - Drop
             if (shouldDrop(dispatch)) {
@@ -65,8 +68,9 @@ internal class DispatchRouter(coroutineDispatcher: CoroutineDispatcher,
                 }
                 return@launch
             }
+
             // Dispatch Send
-            val queue = dequeue(dispatch)
+            val queue = dequeue(dispatch).sortedBy { d -> d.timestamp }
             sendDispatches(queue)
         }
     }
@@ -84,6 +88,19 @@ internal class DispatchRouter(coroutineDispatcher: CoroutineDispatcher,
             }
         }
         return data
+    }
+
+    /**
+     *  Calls the [Transformer.transform] method of each [Transformer] to enrich the [Dispatch]
+     */
+    suspend fun transform(dispatch: Dispatch) {
+        transformers.filter { it.enabled }.forEach {
+            try {
+                it.transform(dispatch)
+            } catch (ex: Exception) {
+                Logger.dev(BuildConfig.TAG, "Failed to transform data from ${it.name}")
+            }
+        }
     }
 
     /**