+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Change Log
+<<<<<<< HEAD
 - 0.1.0 Beta Release: August 2020
 	- Tealium Core
 	- Support for the following: 
@@ -10,3 +11,17 @@
 		- Visitor Service
 		- Remote Commands
 		- Hosted Data Layer
+=======
+
+- 0.1.0 Beta Release: August 2020
+  - Tealium Core
+  - Support for the following modules: 
+    - Tag Management Dispatcher 
+    - Collect Dispatcher
+    - Install Referrer 
+    - Location
+    - Lifecycle Tracking
+    - Visitor Service
+    - Remote Commands
+    - Hosted Data Layer
+>>>>>>> master